Introduce base doc structure and "integration"
We create a basic structure for the documentation of MetalK8s.

The "main" docs are located under the `docs/` directory, and use
reStructuredText syntax.
We added a `docs/solutions/` subdirectory to store all documentation
related to integration of Solutions with MetalK8s.

We also created an `INTEGRATING.md` document at the root of the
repository, to sum up the "integration" docs for quick consumption.
This document is referenced in the root README, so users will easily
navigate to the relevant documentation pages.
Its goal is to provide a quick overview of how MetalK8s wants to
integrate with "Solutions", and further link to "real" documentation.

Issue: GH-951