salt: fix nginx configuration
Rename the new configuration as `.inc` instead of `.conf` to avoid them
to be automagically included by the `/etc/nginx/nginx.conf` in the
container (it contains `include /etc/nginx/conf.d/*.conf;`) which result
in error (because `set` and `location` are then included out of a
`server` block…).

And make sure that the configuration and the images are mounted in the
container.

Also fix the content of 90-metalk8s-registry-config.inc.

Refs: #1049
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>